+++ cmake/SetupCompilerOptions.cmake
@@ -42,6 +42,10 @@ else()
         set(C_COMPILER_FAMILY_IS_INTEL 1)
         message(STATUS "C Compiler family is Intel")
 
+    elseif("${CMAKE_CXX_COMPILER_ID}" STREQUAL "IntelLLVM")
+        set(C_COMPILER_FAMILY_IS_INTELLLVM 1)
+        message(STATUS "C Compiler family is IntelLLVM")
+
     elseif("${CMAKE_CXX_COMPILER_ID}" STREQUAL "PGI")
         set(C_COMPILER_FAMILY_IS_PGI 1)
         message(STATUS "C Compiler family is PGI")
@@ -69,6 +73,10 @@ else()
     elseif("${CMAKE_Fortran_COMPILER_ID}" STREQUAL "Intel")
         set(Fortran_COMPILER_FAMILY_IS_INTEL 1)
         message(STATUS "Fortran Compiler family is Intel")
+
+    elseif("${CMAKE_Fortran_COMPILER_ID}" STREQUAL "IntelLLVM")
+        set(Fortran_COMPILER_FAMILY_IS_INTELLLVM 1)
+        message(STATUS "Fortran Compiler family is IntelLLVM")
 
     elseif("${CMAKE_Fortran_COMPILER_ID}" STREQUAL "PGI")
         set(Fortran_COMPILER_FAMILY_IS_PGI 1)